Several improvements and fixes for ClientPeoplePickerWebServiceInterface implementation
